refactored load pregnancy definition and csv commands
